+++ processor/src/test/resources/bad_input/com/example/observable_ref/NoObservableModel.java
@@ -1,0 +1,16 @@
+package com.example.observable_ref;
+
+import javax.annotation.Nonnull;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.ObservableRef;
+
+@ArezComponent
+public class NoObservableModel
+{
+  @Nonnull
+  @ObservableRef
+  public org.realityforge.arez.Observable getTimeObservable()
+  {
+    throw new IllegalStateException();
+  }
+}