+++ processor/src/test/resources/expected/com/example/observable_ref/Arez_DefaultRefNameModel.java
@@ -1,0 +1,100 @@
+package com.example.observable_ref;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import org.realityforge.arez.Arez;
+import org.realityforge.arez.ArezContext;
+import org.realityforge.arez.Disposable;
+import org.realityforge.arez.Observable;
+
+@Generated("org.realityforge.arez.processor.ArezProcessor")
+public final class Arez_DefaultRefNameModel extends DefaultRefNameModel implements Disposable {
+  private static volatile long $$arez$$_nextId;
+
+  private final long $$arez$$_id;
+
+  private boolean $$arez$$_disposed;
+
+  @Nonnull
+  private final ArezContext $$arez$$_context;
+
+  @Nonnull
+  private final Observable $$arez$$_time;
+
+  public Arez_DefaultRefNameModel() {
+    super();
+    this.$$arez$$_context = Arez.context();
+    this.$$arez$$_id = $$arez$$_nextId++;
+    this.$$arez$$_time = this.$$arez$$_context.createObservable( this.$$arez$$_context.areNamesEnabled() ? $$arez$$_name() + ".time" : null );
+  }
+
+  final long $$arez$$_id() {
+    return $$arez$$_id;
+  }
+
+  String $$arez$$_name() {
+    return "DefaultRefNameModel." + $$arez$$_id();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return $$arez$$_disposed;
+  }
+
+  @Override
+  public void dispose() {
+    if ( !isDisposed() ) {
+      $$arez$$_disposed = true;
+      $$arez$$_time.dispose();
+    }
+  }
+
+  @Override
+  public long getTime() {
+    assert !$$arez$$_disposed;
+    this.$$arez$$_time.reportObserved();
+    return super.getTime();
+  }
+
+  @Override
+  public void setTime(final long time) {
+    assert !$$arez$$_disposed;
+    if ( time != super.getTime() ) {
+      super.setTime(time);
+      this.$$arez$$_time.reportChanged();
+    }
+  }
+
+  @Nonnull
+  @Override
+  public Observable getTimeObservable() {
+    assert !$$arez$$_disposed;
+    return $$arez$$_time;
+  }
+
+  @Override
+  public final int hashCode() {
+    return Long.hashCode( $$arez$$_id() );
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( this == o ) {
+      return true;
+    } else if ( null == o || !(o instanceof Arez_DefaultRefNameModel) ) {
+      return false;
+    } else {
+      final Arez_DefaultRefNameModel that = (Arez_DefaultRefNameModel) o;;
+      return $$arez$$_id() == that.$$arez$$_id();
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( $$arez$$_context.areNamesEnabled() ) {
+      return "ArezComponent[" + $$arez$$_name() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}