+++ processor/src/test/resources/input/com/example/observable_ref/NonStandardNameModel.java
@@ -1,0 +1,27 @@
+package com.example.observable_ref;
+
+import javax.annotation.Nonnull;
+import org.realityforge.arez.annotations.ArezComponent;
+import org.realityforge.arez.annotations.Observable;
+import org.realityforge.arez.annotations.ObservableRef;
+
+@ArezComponent
+public class NonStandardNameModel
+{
+  @Observable
+  public long getTime()
+  {
+    return 0;
+  }
+
+  public void setTime( final long time )
+  {
+  }
+
+  @Nonnull
+  @ObservableRef( name = "time" )
+  public org.realityforge.arez.Observable timeObservable()
+  {
+    throw new IllegalStateException();
+  }
+}